fix view stack bug
